backend/requirements.txt
# Core dependencies
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
WTForms
python-dotenv
Werkzeug
email-validator
cryptography
pymysql

# Reporting (Mandatory for some features)
reportlab

# Analytics (Optional - will be used if installed)
pandas
numpy
plotly
matplotlib